add invoice functionality

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product');
 const { validationResult } = require('express-validator');
+const fileHelper = require('../util/file');
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
@@ -125,6 +126,7 @@ exports.postEditProduct = (req, res, next) => {
     product.price = updatedPrice;
     product.description = updatedDescription;
     if(updatedImage){
+      fileHelper.deleteFile(product.imageUrl);
       product.imageUrl = updatedImage.path;
     }
     
@@ -144,7 +146,16 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const productId = req.body.productId;
-  Product.deleteOne({_id: productId, userId: req.user._id})
+
+  Product.findById(productId)
+  .then(product => {
+    if(!product) {
+      return next(new Error('Product not found!'));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+
+    return Product.deleteOne({_id: productId, userId: req.user._id})
+  })
   .then(() => {
     console.log('Destroyed product');
     res.redirect('/admin/products');
